# Conflicts:
#	src/main/java/com/book/book/BookApplication.java
#	src/main/resources/application.yml

+++ src/main/java/com/book/book/model/vo/UserVO.java
@@ -8,16 +8,16 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserVO {
-    private long id; //序号
+    private Integer id; //序号
     private String nickname; //昵称
     private String username; //用户名
     private String birthday; //生日
     private String tel; //电话
     private String email; //邮箱
     private String address; //地址
-    private long size; //可借数量
-    private long identity; //身份
-    private long isAdmin;
+    private Integer size; //可借数量
+    private Integer identity; //身份
+    private Integer isAdmin;
     @Override
     public String toString() {
         return "UserVO{" +
